requirements.txt
langchain
streamlit
openai
PyPDF2
python-dotenv
tiktoken
faiss-gpu